requirements.txt
langchain
langchain-community
langchain_text_splitters
langchain_openai
langchain_chroma
python_dotenv
langchain-huggingface